mpi/group: split MPIR implementations from MPI functions
