+++ src/mpi/group/group_difference.c
@@ -24,79 +24,6 @@ int MPI_Group_difference(MPI_Group group1, MPI_Group group2, MPI_Group * newgrou
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Group_difference
 #define MPI_Group_difference PMPI_Group_difference
-
-int MPIR_Group_difference_impl(MPIR_Group * group_ptr1, MPIR_Group * group_ptr2,
-                               MPIR_Group ** new_group_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int size1, i, k, g1_idx, g2_idx, l1_pid, l2_pid, nnew;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPIR_GROUP_DIFFERENCE_IMPL);
-
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPIR_GROUP_DIFFERENCE_IMPL);
-    /* Return a group consisting of the members of group1 that are *not*
-     * in group2 */
-    size1 = group_ptr1->size;
-    /* Insure that the lpid lists are setup */
-    MPIR_Group_setup_lpid_pairs(group_ptr1, group_ptr2);
-
-    int *flags = MPL_calloc(size1, sizeof(int), MPL_MEM_OTHER);
-
-    g1_idx = group_ptr1->idx_of_first_lpid;
-    g2_idx = group_ptr2->idx_of_first_lpid;
-
-    nnew = size1;
-    while (g1_idx >= 0 && g2_idx >= 0) {
-        l1_pid = group_ptr1->lrank_to_lpid[g1_idx].lpid;
-        l2_pid = group_ptr2->lrank_to_lpid[g2_idx].lpid;
-        if (l1_pid < l2_pid) {
-            g1_idx = group_ptr1->lrank_to_lpid[g1_idx].next_lpid;
-        } else if (l1_pid > l2_pid) {
-            g2_idx = group_ptr2->lrank_to_lpid[g2_idx].next_lpid;
-        } else {
-            /* Equal */
-            flags[g1_idx] = 1;
-            g1_idx = group_ptr1->lrank_to_lpid[g1_idx].next_lpid;
-            g2_idx = group_ptr2->lrank_to_lpid[g2_idx].next_lpid;
-            nnew--;
-        }
-    }
-    /* Create the group */
-    if (nnew == 0) {
-        /* See 5.3.2, Group Constructors.  For many group routines,
-         * the standard explicitly says to return MPI_GROUP_EMPTY;
-         * for others it is implied */
-        *new_group_ptr = MPIR_Group_empty;
-        goto fn_exit;
-    } else {
-        mpi_errno = MPIR_Group_create(nnew, new_group_ptr);
-        /* --BEGIN ERROR HANDLING-- */
-        if (mpi_errno) {
-            goto fn_fail;
-        }
-        /* --END ERROR HANDLING-- */
-        (*new_group_ptr)->rank = MPI_UNDEFINED;
-        k = 0;
-        for (i = 0; i < size1; i++) {
-            if (!flags[i]) {
-                (*new_group_ptr)->lrank_to_lpid[k].lpid = group_ptr1->lrank_to_lpid[i].lpid;
-                if (i == group_ptr1->rank)
-                    (*new_group_ptr)->rank = k;
-                k++;
-            }
-        }
-        /* TODO calculate is_local_dense_monotonic */
-    }
-
-    MPL_free(flags);
-
-  fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPIR_GROUP_DIFFERENCE_IMPL);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-
 #endif
