+++ src/mpi/group/group_range_excl.c
@@ -24,89 +24,6 @@ int MPI_Group_range_excl(MPI_Group group, int n, int ranges[][3], MPI_Group * ne
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Group_range_excl
 #define MPI_Group_range_excl PMPI_Group_range_excl
-int MPIR_Group_range_excl_impl(MPIR_Group * group_ptr, int n, int ranges[][3],
-                               MPIR_Group ** new_group_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int size, i, j, k, nnew, first, last, stride;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPIR_GROUP_RANGE_EXCL_IMPL);
-
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPIR_GROUP_RANGE_EXCL_IMPL);
-    /* Compute size, assuming that included ranks are valid (and distinct) */
-    size = group_ptr->size;
-    nnew = 0;
-    for (i = 0; i < n; i++) {
-        first = ranges[i][0];
-        last = ranges[i][1];
-        stride = ranges[i][2];
-        /* works for stride of either sign.  Error checking above
-         * has already guaranteed stride != 0 */
-        nnew += 1 + (last - first) / stride;
-    }
-    nnew = size - nnew;
-
-    if (nnew == 0) {
-        *new_group_ptr = MPIR_Group_empty;
-        goto fn_exit;
-    }
-
-    /* Allocate a new group and lrank_to_lpid array */
-    mpi_errno = MPIR_Group_create(nnew, new_group_ptr);
-    /* --BEGIN ERROR HANDLING-- */
-    if (mpi_errno) {
-        goto fn_fail;
-    }
-    /* --END ERROR HANDLING-- */
-    (*new_group_ptr)->rank = MPI_UNDEFINED;
-
-    /* Group members are taken in rank order from the original group,
-     * with the specified members removed. Use the flag array for that
-     * purpose.  If this was a critical routine, we could use the
-     * flag values set in the error checking part, if the error checking
-     * was enabled *and* we are not MPI_THREAD_MULTIPLE, but since this
-     * is a low-usage routine, we haven't taken that optimization.  */
-
-    int *flags = MPL_calloc(size, sizeof(int), MPL_MEM_OTHER);
-
-    for (i = 0; i < n; i++) {
-        first = ranges[i][0];
-        last = ranges[i][1];
-        stride = ranges[i][2];
-        if (stride > 0) {
-            for (j = first; j <= last; j += stride) {
-                flags[j] = 1;
-            }
-        } else {
-            for (j = first; j >= last; j += stride) {
-                flags[j] = 1;
-            }
-        }
-    }
-    /* Now, run through the group and pick up the members that were
-     * not excluded */
-    k = 0;
-    for (i = 0; i < size; i++) {
-        if (!flags[i]) {
-            (*new_group_ptr)->lrank_to_lpid[k].lpid = group_ptr->lrank_to_lpid[i].lpid;
-            if (group_ptr->rank == i) {
-                (*new_group_ptr)->rank = k;
-            }
-            k++;
-        }
-    }
-
-    MPL_free(flags);
-
-    /* TODO calculate is_local_dense_monotonic */
-
-
-  fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPIR_GROUP_RANGE_EXCL_IMPL);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif
