+++ src/mpi/group/group_translate_ranks.c
@@ -25,80 +25,6 @@ int MPI_Group_translate_ranks(MPI_Group group1, int n, const int ranks1[], MPI_G
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Group_translate_ranks
 #define MPI_Group_translate_ranks PMPI_Group_translate_ranks
-
-int MPIR_Group_translate_ranks_impl(MPIR_Group * gp1, int n, const int ranks1[],
-                                    MPIR_Group * gp2, int ranks2[])
-{
-    int mpi_errno = MPI_SUCCESS;
-    int i, g2_idx, l1_pid, l2_pid;
-
-    MPL_DBG_MSG_S(MPIR_DBG_OTHER, VERBOSE, "gp2->is_local_dense_monotonic=%s",
-                  (gp2->is_local_dense_monotonic ? "TRUE" : "FALSE"));
-
-    /* Initialize the output ranks */
-    for (i = 0; i < n; i++)
-        ranks2[i] = MPI_UNDEFINED;
-
-    if (gp2->size > 0 && gp2->is_local_dense_monotonic) {
-        /* g2 probably == group_of(MPI_COMM_WORLD); use fast, constant-time lookup */
-        int lpid_offset = gp2->lrank_to_lpid[0].lpid;
-
-        MPIR_Assert(lpid_offset >= 0);
-        for (i = 0; i < n; ++i) {
-            int g1_lpid;
-
-            if (ranks1[i] == MPI_PROC_NULL) {
-                ranks2[i] = MPI_PROC_NULL;
-                continue;
-            }
-            /* "adjusted" lpid from g1 */
-            g1_lpid = gp1->lrank_to_lpid[ranks1[i]].lpid - lpid_offset;
-            if ((g1_lpid >= 0) && (g1_lpid < gp2->size)) {
-                ranks2[i] = g1_lpid;
-            }
-            /* else leave UNDEFINED */
-        }
-    } else {
-        /* general, slow path; lookup time is dependent on the user-provided rank values! */
-        g2_idx = gp2->idx_of_first_lpid;
-        if (g2_idx < 0) {
-            MPII_Group_setup_lpid_list(gp2);
-            g2_idx = gp2->idx_of_first_lpid;
-        }
-        if (g2_idx >= 0) {
-            /* g2_idx can be < 0 if the g2 group is empty */
-            l2_pid = gp2->lrank_to_lpid[g2_idx].lpid;
-            for (i = 0; i < n; i++) {
-                if (ranks1[i] == MPI_PROC_NULL) {
-                    ranks2[i] = MPI_PROC_NULL;
-                    continue;
-                }
-                l1_pid = gp1->lrank_to_lpid[ranks1[i]].lpid;
-                /* Search for this l1_pid in group2.  Use the following
-                 * optimization: start from the last position in the lpid list
-                 * if possible.  A more sophisticated version could use a
-                 * tree based or even hashed search to speed the translation. */
-                if (l1_pid < l2_pid || g2_idx < 0) {
-                    /* Start over from the beginning */
-                    g2_idx = gp2->idx_of_first_lpid;
-                    l2_pid = gp2->lrank_to_lpid[g2_idx].lpid;
-                }
-                while (g2_idx >= 0 && l1_pid > l2_pid) {
-                    g2_idx = gp2->lrank_to_lpid[g2_idx].next_lpid;
-                    if (g2_idx >= 0)
-                        l2_pid = gp2->lrank_to_lpid[g2_idx].lpid;
-                    else
-                        l2_pid = -1;
-                }
-                if (l1_pid == l2_pid)
-                    ranks2[i] = g2_idx;
-            }
-        }
-    }
-    return mpi_errno;
-}
-
-
 #endif
 
 /*@
