+++ src/mpi/group/group_union.c
@@ -24,114 +24,6 @@ int MPI_Group_union(MPI_Group group1, MPI_Group group2, MPI_Group * newgroup)
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Group_union
 #define MPI_Group_union PMPI_Group_union
-
-int MPIR_Group_union_impl(MPIR_Group * group_ptr1, MPIR_Group * group_ptr2,
-                          MPIR_Group ** new_group_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int g1_idx, g2_idx, nnew, i, k, size1, size2, mylpid;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPIR_GROUP_UNION_IMPL);
-
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPIR_GROUP_UNION_IMPL);
-
-    /* Determine the size of the new group.  The new group consists of all
-     * members of group1 plus the members of group2 that are not in group1.
-     */
-    g1_idx = group_ptr1->idx_of_first_lpid;
-    g2_idx = group_ptr2->idx_of_first_lpid;
-
-    /* If the lpid list hasn't been created, do it now */
-    if (g1_idx < 0) {
-        MPII_Group_setup_lpid_list(group_ptr1);
-        g1_idx = group_ptr1->idx_of_first_lpid;
-    }
-    if (g2_idx < 0) {
-        MPII_Group_setup_lpid_list(group_ptr2);
-        g2_idx = group_ptr2->idx_of_first_lpid;
-    }
-    nnew = group_ptr1->size;
-
-    /* Clear the flag bits on the second group.  The flag is set if
-     * a member of the second group belongs to the union */
-    size2 = group_ptr2->size;
-    int *flags = MPL_calloc(size2, sizeof(int), MPL_MEM_OTHER);
-
-    /* Loop through the lists that are ordered by lpid (local process
-     * id) to detect which processes in group 2 are not in group 1
-     */
-    while (g1_idx >= 0 && g2_idx >= 0) {
-        int l1_pid, l2_pid;
-        l1_pid = group_ptr1->lrank_to_lpid[g1_idx].lpid;
-        l2_pid = group_ptr2->lrank_to_lpid[g2_idx].lpid;
-        if (l1_pid > l2_pid) {
-            nnew++;
-            flags[g2_idx] = 1;
-            g2_idx = group_ptr2->lrank_to_lpid[g2_idx].next_lpid;
-        } else if (l1_pid == l2_pid) {
-            g1_idx = group_ptr1->lrank_to_lpid[g1_idx].next_lpid;
-            g2_idx = group_ptr2->lrank_to_lpid[g2_idx].next_lpid;
-        } else {
-            /* l1 < l2 */
-            g1_idx = group_ptr1->lrank_to_lpid[g1_idx].next_lpid;
-        }
-    }
-    /* If we hit the end of group1, add the remaining members of group 2 */
-    while (g2_idx >= 0) {
-        nnew++;
-        flags[g2_idx] = 1;
-        g2_idx = group_ptr2->lrank_to_lpid[g2_idx].next_lpid;
-    }
-
-    if (nnew == 0) {
-        *new_group_ptr = MPIR_Group_empty;
-        goto fn_exit;
-    }
-
-    /* Allocate a new group and lrank_to_lpid array */
-    mpi_errno = MPIR_Group_create(nnew, new_group_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    /* If this process is in group1, then we can set the rank now.
-     * If we are not in this group, this assignment will set the
-     * current rank to MPI_UNDEFINED */
-    (*new_group_ptr)->rank = group_ptr1->rank;
-
-    /* Add group1 */
-    size1 = group_ptr1->size;
-    for (i = 0; i < size1; i++) {
-        (*new_group_ptr)->lrank_to_lpid[i].lpid = group_ptr1->lrank_to_lpid[i].lpid;
-    }
-
-    /* Add members of group2 that are not in group 1 */
-
-    if (group_ptr1->rank == MPI_UNDEFINED && group_ptr2->rank >= 0) {
-        mylpid = group_ptr2->lrank_to_lpid[group_ptr2->rank].lpid;
-    } else {
-        mylpid = -2;
-    }
-    k = size1;
-    for (i = 0; i < size2; i++) {
-        if (flags[i]) {
-            (*new_group_ptr)->lrank_to_lpid[k].lpid = group_ptr2->lrank_to_lpid[i].lpid;
-            if ((*new_group_ptr)->rank == MPI_UNDEFINED &&
-                group_ptr2->lrank_to_lpid[i].lpid == mylpid)
-                (*new_group_ptr)->rank = k;
-            k++;
-        }
-    }
-
-    MPL_free(flags);
-
-    /* TODO calculate is_local_dense_monotonic */
-
-  fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPIR_GROUP_UNION_IMPL);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-
 #endif
